Stabilized System by Fixing a Unit Conversion
While trying to understand what accelerated the system, I realized that the system is multiplied by around 7.6 every iteration. I narrowed my search down to the conversion constant c and how it is used in line 114/115. I changed it to have that value be multiplied by c, and now the system is stable.